pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-ascii"
version = "0.1.0"
description = "A powerful Python package to convert images and videos to highly customizable ASCII art."
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
    { name = "PluvioXO" }
]
keywords = ["ascii", "art", "image", "video", "converter", "terminal", "generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "Pillow>=9.0.0",
    "numpy>=2.0.0",
    "opencv-python>=4.5.0"
]

[project.urls]
Homepage = "https://github.com/PluvioXO/auto-ascii"
Repository = "https://github.com/PluvioXO/auto-ascii"

[project.scripts]
auto-ascii = "auto_ascii.cli:main"
